HID: appleib: Add Apple iBridge HID driver for T1 chip
The iBridge device provides access to several devices, including:
- the Touch Bar
- the iSight webcam
- the light sensor
- the fingerprint sensor

This driver provides the core support for managing the iBridge device
and the access to the underlying devices. In particular, the
functionality for the Touch Bar and ALs is exposed via USB HID
interfaces, and on devices with the T1 chip one of the HID devices is
used for both functions. So this driver creates virtual HID devices, one
per top-level report collection on the HID device (for a total of 2
virtual HID devices). The sub-drivers then bind to these virtual HID
devices.

This way the Touch Bar backlight and ALS drivers can be kept in their
own modules, while at the same time making them look very much like as
if they were connected to the real HID devices. And those drivers then
work (mostly) without further changes on MacBooks with the T2 chip that
don't need this driver.

This driver is largely based on previous work done by Ronald Tschalär
<ronald@innovation.ch>.

Signed-off-by: Kerem Karabay <kekrby@gmail.com>