HID: core: Add a quirk to force report scanning for devices that have a special driver
The Apple iBridge device exposes multiple USB HID interfaces, some of
which need to be handled by special drivers, while others need to be
handled by hid-multitouch and hid-sensor-hub, which use the group system
to determine which devices to bind to. Currently, the HID core doesn't
scan the reports on devices that have the HID_QUIRK_HAS_SPECIAL_DRIVER
quirk set, so the groups of the ALS and digitizer interfaces don't get
set and the drivers don't bind to them. Additionally, the upcoming
hid-appleib driver will also make use of the group mechanism, so add a
quirk to force report scanning.

Signed-off-by: Kerem Karabay <kekrby@gmail.com>